logic implemented
