+++ src/commands/farm.rs
@@ -1,11 +1,11 @@
 use color_eyre::eyre::Result;
 use color_eyre::eyre::WrapErr;
+use subspace_sdk::Farmer;
 use subspace_sdk::{chain_spec, Node, PlotDescription, PublicKey};
-use subspace_sdk::{Farmer, NodeMode};
 use tracing::instrument;
 
 use crate::config::parse_config;
-use crate::utils::node_directory_getter;
+use crate::utils::{install_tracing, node_directory_getter};
 
 pub(crate) struct FarmingArgs {
     reward_address: PublicKey,
@@ -13,9 +13,12 @@ pub(crate) struct FarmingArgs {
     plot: PlotDescription,
 }
 
-pub(crate) async fn farm() -> Result<()> {
+pub(crate) async fn farm(is_verbose: bool) -> Result<()> {
+    install_tracing(is_verbose);
+    color_eyre::install()?;
     let args = prepare_farming().await?;
     start_farming(args).await?;
+
     Ok(())
 }
 
@@ -47,7 +50,6 @@ async fn prepare_farming() -> Result<FarmingArgs> {
     };
 
     let node: Node = Node::builder()
-        .mode(NodeMode::Full)
         .name(node_name)
         .build(node_directory, chain)
         .await
